Complete Course Upload 1
This is a major update to the repository that includes virtually all of the course material, the assignments, the model answers. There are some exempted data, unfortunately. In particular, we have not included a tweets database.

Future versions of this course will adapt to provide fully open data that still demonstrates the same skills. In the meantime this is for perusal.